Neb Dynarec, part 1: Instruction simplification and cached interpretation
Instruction simplification transforms seemingly useless B<cond>
instructions that simply skip their delay slot (offset +1) into NOPs,
as well as transforming 32-bit and 64-bit register moves, register
zeroing and small constant loading into ADDUs, ORs and ORIs.

This simplification hopefully increases the number of successive N64
instructions that can run solely using functions that are still cached
by the host's instruction cache.

If the NEB_DYNAREC define is used, the Neb Dynarec is used when the
configuration calls for using emulation core #2, on amd64 and mipsel32.
Currently it only performs cached interpretation, but this will change.

Some opcode emitters are present in this commit, but are not used. MIPS
emitters are present for all operations, and AMD64 emitters are present
only for some. JIT code will be added in a later commit.

The Neb Dynarec introduces its own opcode functions, which are slightly
different from those of cached_interp.c, and this is all so that the
code that handles exceptions in the delay slot of a branch can be
massively simplified:
In that case, instead of setting next_interupt (the cycles to next
interrupt counter) to a sentinel (0), and relying on the jump-opcode
function to always perform this code...

  if (next_interupt <= Count) gen_interupt();
  // and gen_interupt will see that next_interupt == 0, and set the
  // PC to the exception handler (!) then recalculate next_interupt
  // (wasteful!!)

... before returning, we instead unset the delay_slot flag, set the PC
to service the pending exception, and exit the opcode function if the
skip_jump variable is non-zero. This will also massively simplify the
code that needs to be emitted in the JIT later, because it will not
need to contain special cases to unconditionally execute similar code
to service a pseudo-interrupt. It can simply return after unsetting the
skip_jump variable, because delay_slot is cleared by exception.c if it
was set.

Because this behavior is unexpected by the other JITs and by the other
interpreters and would break them, it is only enabled in exception.c if
NEB_DYNAREC is defined.