Neb Dynarec, part 2: Minimal IL and architecture emitters for NOP
An intermediate language to represent MIPS opcodes and the transfer of
values between instructions in a block has been added. It is currently
sufficient for NOP functions but is intended to be extended with more
instructions.

The driver passes the range of instructions having enough emitters to
make a function to the IL emitter, which currently adds only 4 IL
instructions to the function. It is intended to be used for constant
propagation once more opcodes are added to it.

Following this, the IL is sent to the architecture instruction emitter.
Currently, no register allocation is performed for the registers needed
by IL_OP_SET_PC (load a native register with &PC, then another with the
address of a precomp_instr). It will come later.

Because of the 'avail' (bytes available) parameter in arch_emit_code,
opcode helpers can now help it track how much bytes are left without
any heuristics. These opcode helpers were left over from an earlier
incarnation of the Neb Dynarec and didn't do this sort of tracking,
instead relying on checking before a block was emitted whether the
largest possible block on the architecture would fit. Opcode helpers
are modified to accept a pointer to arch_emit_code's 'avail' parameter.

mipsel32:
* Rename FlushCodeCache to flush_icache to properly represent its role,
  which is to flush the instruction cache and not the Neb Dynarec's
  code cache.
* Correct flush_icache's behaviour when compiled for processors not
  having the SYNCI instruction (MIPS32R2).
* Add enough emitters for IL_OP_ENTER_FRAME, IL_OP_EXIT_FRAME (dummy),
  IL_OP_SET_PC and IL_OP_RETURN.

amd64:
* Add caller-saved and callee-saved register counts to config.h.
* Add virtual (virt_reg_t) to native register mappings.
* Add enough emitters for IL_OP_ENTER_FRAME, IL_OP_EXIT_FRAME (dummy),
  IL_OP_SET_PC and IL_OP_RETURN.