Neb Dynarec: Allow compiling one instruction at the start of a page
Formerly, compiling a block at the start of a 4 KiB N64 page was
forbidden. This restriction is lifted, but only for one instruction,
because making a block out of multiple instructions and running it
as the delay slot of a branch possibly ending the previous page is
not allowed by the original N64 code.