LLVM JIT: Recompile in-page BEQ, BNE, BGEZ, BLTZ, BLEZ, BGTZ
The implementations only update last_addr (start of the last run of
instructions) and Count if the branch succeeds.

In-page branches may also happen to jump to a different place in the
same function being compiled to LLVM IR. In that case, they can jump
directly to the basic block backing the branch target.