LLVM JIT: Recompile in-page BEQL, BNEL, BGEZL, BLTZL, BLEZL, BGTZL
The implementations only update last_addr (start of the last run of
instructions) and Count if the branch succeeds.

In-page branches may also happen to jump to a different place in the
same function being compiled to LLVM IR. In that case, they can jump
directly to the basic block backing the branch target.

Branch Likely instructions only execute their delay slot if they
succeed. As with the interpreters, the delay slot is just not executed
if the branches fail; on real hardware, the delay slot would be
executed but its result would be cancelled.